+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## Unreleased
+
+### Added
+- added table of contents with songs title and arrangement to song sheets
+
 ## 0.10.10 (2025-10-04)
 
 ### Added